fix: add hasil pembahasan

+++ src/features/cbt/statistik-nilai.tsx
@@ -14,15 +14,28 @@ export function StatistikNilai({
   const navigate = useNavigate()
 
   const handleStartExam = (item) => {
-    const startTime = new Date().toISOString()
-    const duration = item?.waktu_ujian
+    // Cek apakah mulaiUjian sudah ada di local storage
+    const existingData = localStorage.getItem('mulaiujian')
+    let startTime = null
 
-    const dataToSave = {
-      startTime,
-      duration,
+    // Jika sudah ada data, gunakan startTime dari data yang sudah ada
+    if (existingData) {
+      const existingStartTime = JSON.parse(existingData).startTime
+      if (existingStartTime) {
+        startTime = existingStartTime
+      }
+    } else {
+      // Jika belum ada data, buat startTime baru
+      startTime = new Date().toISOString()
+      const duration = item?.waktu_ujian
+
+      const dataToSave = {
+        startTime,
+        duration,
+      }
+
+      localStorage.setItem('mulaiujian', JSON.stringify(dataToSave))
     }
-
-    localStorage.setItem('mulaiujian', JSON.stringify(dataToSave))
     navigate(`/exam?soal=${item?.id_ujian}`)
   }
 
@@ -42,9 +55,14 @@ export function StatistikNilai({
         })}
       >
         {/* --- Card Title --- */}
-        <div className="col-span-5 flex flex-col gap-y-12 rounded-2xl bg-slate-50 p-12 phones:col-span-12">
+        <div className="col-span-6 flex flex-col gap-y-12 rounded-2xl bg-slate-50 p-12 phones:col-span-12">
           <div className="flex items-center gap-x-12 text-[1.2rem]">
-            <div className={'rounded-full bg-slate-300 px-16 py-4'}>
+            <div
+              className={clsx('rounded-full px-16 py-4', {
+                'bg-slate-300 ': item?.status === 0,
+                'bg-green-700 text-white': item?.status === 1,
+              })}
+            >
               {item?.skor === 'Belum Mengerjakan Ujian'
                 ? 'Belum Dikerjakan'
                 : 'Sudah Dikerjakan'}
@@ -63,7 +81,7 @@ export function StatistikNilai({
               {dayjs(item?.tanggal_akhir).format('hh:mm')}
             </p>
           </div>
-          {typeof item?.skor !== 'number' && (
+          {item?.status === 0 ? (
             <div
               className="rounded-lg border border-transparent bg-primary py-12 text-center text-[1.4rem] text-white hover:cursor-pointer hover:border-primary hover:bg-transparent hover:text-primary phones:text-[1.6rem]"
               onClick={() => {
@@ -72,52 +90,72 @@ export function StatistikNilai({
             >
               Kerjakan Soal
             </div>
+          ) : (
+            <div
+              className="rounded-lg border border-transparent bg-primary py-12 text-center text-[1.4rem] text-white hover:cursor-pointer hover:border-primary hover:bg-transparent hover:text-primary phones:text-[1.6rem]"
+              onClick={() => {
+                navigate(`/result?soal=${item?.id_ujian}`)
+              }}
+            >
+              Lihat Hasil
+            </div>
           )}
         </div>
-        {typeof item?.skor === 'number' && (
+        {/* Skor  */}
+        {item?.status === 1 && (
           <div className="col-span-3 flex flex-col items-center justify-center gap-y-12 rounded-2xl bg-slate-50 p-12">
             <span>
               <Award size={40} />
             </span>
             <p className="font-medium">
-              Nilai{' '}
-              <span className="font-bold text-primary">{item?.skor} / </span>
-              {item?.max_skor}
+              Skor: <span className="font-bold text-primary">{item?.skor}</span>
             </p>
           </div>
         )}
-        {typeof item?.skor === 'number' && (
-          <div className="col-span-4 flex flex-col items-center justify-center gap-y-12 rounded-2xl bg-slate-50 p-12">
+        {/* Peringkat */}
+        {item?.status === 1 && (
+          <div className="col-span-3 flex flex-col items-center justify-center gap-y-12 rounded-2xl bg-slate-50 p-12">
             <span>
               <Trophy size={40} />
             </span>
             <p className="font-medium">
               Peringkat{' '}
               <span className="font-bold text-primary">
-                {item?.peringkat} /{' '}
+                {item?.peringkat ?? '-'}
               </span>
-              {item?.total_peserta}
             </p>
           </div>
         )}
       </div>
 
-      {typeof item?.skor === 'number' && (
+      {item?.status === 1 && (
         <div className="grid grid-cols-12 gap-16">
           <div className="col-span-6">
-            <p className="mb-8 font-semibold">Total Nilai</p>
-            {item?.nilai?.map((data, id) => (
-              <div
-                className="col-span-6 gap-y-12 rounded-2xl border-b p-12"
-                key={id}
-              >
-                {data?.title}:{' '}
-                <span className="font-bold text-primary">{data?.true}</span>/
-                {data?.total}
-              </div>
-            ))}
+            <p className="mb-8 font-semibold">Deskripsi</p>
+            <div className="col-span-6 gap-y-12 rounded-2xl border-b p-12">
+              Soal Dijawab:{' '}
+              <span className="font-bold text-primary">{item?.dijawab}</span>/
+              {item?.jumlah_soal}
+            </div>
+            <div className="col-span-6 gap-y-12 rounded-2xl border-b p-12">
+              Soal Tidak Dijawab:{' '}
+              <span className="font-bold text-primary">
+                {item?.tidak_dijawab}
+              </span>
+              /{item?.jumlah_soal}
+            </div>
+            <div className="col-span-6 gap-y-12 rounded-2xl border-b p-12">
+              Soal Benar:{' '}
+              <span className="font-bold text-primary">{item?.benar}</span>/
+              {item?.jumlah_soal}
+            </div>
+            <div className="col-span-6 gap-y-12 rounded-2xl border-b p-12">
+              Soal Salah:{' '}
+              <span className="font-bold text-primary">{item?.salah}</span>/
+              {item?.jumlah_soal}
+            </div>
           </div>
-          <div className="col-span-6 flex flex-col gap-y-12">
+          <div className="col-span-6 flex flex-col items-center justify-center gap-y-12">
             <div className="flex flex-col items-center">
               <div
                 className={clsx(
@@ -132,15 +170,15 @@ export function StatistikNilai({
               </div>
             </div>
             <div className="text-center">
-              {Number(item?.skor) > 300 ? (
+              {item?.status_lulus ? (
                 <div className="flex flex-col gap-y-4">
-                  <p>Congratulation!</p>
-                  <p> Your score passed the passsing grade</p>
+                  <p>Selamat!</p>
+                  <p>Skor anda melewati passsing grade</p>
                 </div>
               ) : (
                 <div className="flex flex-col gap-y-4">
-                  <p>Unfortunely!</p>
-                  <p> Your score below the passsing grade</p>
+                  <p>Sayang sekali!</p>
+                  <p>Skor anda dibawah passsing grade</p>
                 </div>
               )}
             </div>
